# src/schemas/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "The Hire Me Project API",
    "version": "1.0.0",
    "description": "Route definitions for THMP API"
  },
...
  "paths": {
...
    "/api/jobs/getAll": {
      "get": {
        "tags": [
          "Job"
        ],
        "summary": "Get all jobs",
        "description": "Get all jobs",
        "responses": {
          "200": {
            "description": "A list of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "description": "Total number of jobs",
                      "example": 10
                    },
                    "jobs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/jobs/searchAndCreate": {
      "post": {
        "tags": [
          "Job"
        ],
        "summary": "Search and create jobs",
        "description": "Searches for jobs based on the keywords and creates them in the database",
        "parameters": [
          {
            "name": "keywords",
            "in": "path",
            "description": "Keywords to search for",
            "required": true,
            "schema": {
              "type": "string",
              "example": "software engineer"
            }
          },
          {
            "name": "locationId",
            "in": "path",
            "description": "Location ID to search for jobs",
            "required": false,
            "schema": {
              "type": "number",
              "example": "104769905 (Sydney, Australia)",
              "default": 104769905
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Job object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Created 25 jobs out of 25"
                }
              }
            }
          },
          "400": {
            "description": "Missing query params"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/jobs/approveByGPT": {
      "patch": {
        "tags": [
          "Job"
        ],
        "summary": "Approve jobs by GPT",
        "description": "Route that checks on jobs that are pending approval by GPT, are easy to apply to, have been approved by the formula. According to an uploaded resume and the job description, the job is approved or rejected by chatgpt model 4 turbo.",
        "responses": {
          "200": {
            "description": "Job object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Jobs approved: 5 out of 10"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Job": {
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "company"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the job"
          },
          "title": {
            "type": "string",
            "description": "The title of the job"
          },
          "url": {
            "type": "string",
            "description": "The url of the job in LinkedIn"
          },
          "referenceId": {
            "type": "string",
            "description": "The reference id of the job"
          },
          "posterId": {
            "type": "string",
            "description": "The poster id of the job"
          },
          "company": {
            "type": "string",
            "description": "The company of the job"
          },
          "location": {
            "type": "string",
            "description": "The string id of the location of the job"
          },
          "type": {
            "type": "string",
            "description": "The type of the job"
          },
          "postDate": {
            "type": "string",
            "description": "The post date of the job"
          },
          "benefits": {
            "type": "string",
            "description": "The benefits of the job"
          },
          "approvedByFormula": {
            "type": "string",
            "enum": [
              "yes",
              "no",
              "pending"
            ],
            "default": "pending",
            "description": "The approved by formula output of the job"
          },
          "approvedByGPT": {
            "type": "string",
            "enum": [
              "yes",
              "no",
              "pending"
            ],
            "default": "pending",
            "description": "The approved by OpenAI chat completion of the job"
          },
          "easyApply": {
            "type": "string",
            "enum": [
              "yes",
              "no",
              "pending"
            ],
            "default": "pending",
            "description": "If the job has an easy apply option"
          },
          "Keywords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Keyword"
            },
            "description": "The keywords of the job"
          }
        }
      },
...
      "Keyword": {
        "type": "object",
        "required": [
          "id",
          "keyword"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "keyword": {
            "type": "string",
            "example": "Software Engineer"
          }
        }
      }
    }
  }
}